♻️ refactor: 코드 린트 수정
