+++ air-quality-app-batch/src/main/java/io/howstheairtoday/airqualityappbatch/listener/JobCompletionNotificationListener.java
@@ -1,4 +1,4 @@
-package io.howstheairtoday.airqualityappbatch.batch.listener;
+package io.howstheairtoday.airqualityappbatch.listener;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -15,7 +15,7 @@ public class JobCompletionNotificationListener implements JobExecutionListener {
 
     @Override
     public void afterJob(JobExecution jobExecution) {
-        if(jobExecution.getStatus() == BatchStatus.COMPLETED) {
+        if (jobExecution.getStatus() == BatchStatus.COMPLETED) {
             log.info("!!! JOB FINISHED! Time to verify the results");
         }
     }
